1. AI and ML Basics/9. Document GPT Chatboat/requirements.txt
streamlit
openai
langchain
faiss-cpu
python-docx
PyPDF2
tiktoken
